Move sessions related classes to the internal package
COSDK-500
