+++ core/api/core.api
@@ -927,20 +927,6 @@ public final class com/adyen/checkout/core/sessions/SessionModel$Creator : andro
 	public synthetic fun newArray (I)[Ljava/lang/Object;
 }
 
-public final class com/adyen/checkout/core/sessions/SessionSavedStateHandleContainer : com/adyen/checkout/core/common/internal/SavedStateHandleContainer {
-	public static final field $stable I
-	public static final field Companion Lcom/adyen/checkout/core/sessions/SessionSavedStateHandleContainer$Companion;
-	public fun <init> (Landroidx/lifecycle/SavedStateHandle;Lcom/adyen/checkout/core/sessions/CheckoutSession;)V
-	public fun getSavedStateHandle ()Landroidx/lifecycle/SavedStateHandle;
-	public final fun getSessionModel ()Lcom/adyen/checkout/core/sessions/SessionModel;
-	public final fun isFlowTakenOver ()Ljava/lang/Boolean;
-	public final fun setFlowTakenOver (Ljava/lang/Boolean;)V
-	public final fun updateSessionData (Ljava/lang/String;)V
-}
-
-public final class com/adyen/checkout/core/sessions/SessionSavedStateHandleContainer$Companion {
-}
-
 public final class com/adyen/checkout/core/sessions/SessionSetupConfiguration : com/adyen/checkout/core/common/internal/model/ModelObject {
 	public static final field $stable I
 	public static final field CREATOR Landroid/os/Parcelable$Creator;
